Update digInAReviewOfIntegration.tex
Fixed some errors.

Line 59:  a \answer{\frac{1}{\sqrt{a^2-x^2}}} ---> \frac{a}{\sqrt{\answer{a^2-x^2}}} to prevent branch cut errors.  I answered the question correctly, but answer was not accepted because it passed to a different branch of sqrt.  This rephrasing avoids that.

Line 183:  Had the incorrect solution before.